+++ src/component/Duy.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+function Duy(){
+    return(
+        <div>Nguyen Ngoc Duy</div>
+    )
+}
+export default Duy
